# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Baseline API
  version: 0.1.0
  description: >
    Thin longevity backend. Whoop arrives via OAuth + webhooks (server-side);
    Apple Watch / Android arrives via the device shim posting daily aggregates.
    Every read endpoint reports values against the user's own 30-day baseline.
paths:
...
  /v1/annotations:
    get:
      summary: List context tags
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Annotation" }
    post:
      summary: Tag a day (travel, illness, alcohol, ...)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Annotation" }
      responses: { "200": { description: Created } }
  /v1/ingest/samples:
    post:
      summary: Device shim posts pre-aggregated daily values (HealthKit / Health Connect)
      security: [ { deviceToken: [] } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                batch_id: { type: string }
                samples:
                  type: array
                  items: { $ref: "#/components/schemas/DeviceSample" }
      responses: { "200": { description: Accepted } }
...
components:
  securitySchemes:
    deviceToken: { type: http, scheme: bearer }
  schemas:
...
    Annotation:
      type: object
      required: [day, tag]
      properties:
        day: { type: string, format: date }
        tag: { type: string }
        note: { type: string }
    DeviceSample:
      type: object
      required: [day, metric, value]
      properties:
        day: { type: string, format: date }
        metric: { type: string, enum: [steps, active_kcal, vo2max, hrv_sdnn_ms, hrv_rmssd_ms, resting_hr, sleep_min] }
        value: { type: number }